# Peken-Banyumasan-Gate-Masuk-Keluar/openapi-gate.yaml
openapi: 3.1.0
info:
  title: Peken Banyumasan — Gate (Admin & Petugas) API
  version: 2.4.1
  description: |
    API consumed by the Gate admin/petugas panel
    (`Peken-Banyumasan-Gate-Masuk-Keluar/frontend/`).

    **Auth (v2.0 — Supabase):**
    - Login/logout/me/changePassword are **Supabase-direct** — frontend calls
      `supabase.auth.signInWithPassword()` etc. without touching this backend.
    - Gate has **no self-register** — admin and petugas accounts are created by superadmin.
    - `PUT /api/auth/profile` (custom fields only) is **HYBRID**.
    - All other endpoints require a valid **Supabase JWT** in `Authorization: Bearer <token>`.
      Backend verifies using `SUPABASE_JWT_SECRET` and enforces `app_metadata.role IN ('admin', 'petugas')`.
    - Endpoints exclusive to `role: admin` are noted in their summary. Petugas is limited to
      dashboard, visitor scan, and basic event/zone read.

    **Vocabulary (CANONICAL — non-negotiable):**
    - `kolaborator` (NEVER member, kreator)
    - `artisan` (NEVER tenant, umkm)
    - `aktivitas` for admin-moderation story feed (NEVER admin/stories)
    - Path segment `/kolaborator` singular (NEVER `/kolaborators`)
    - Path segment `/artisan` singular (NEVER `/artisans`)

    ---

    **Response Envelope (ALL endpoints except binary file downloads):**
    ```
    Success: { "status": "success", "message": "<optional>", "data": <payload> }
...
    ```
...

    **Authentication:**
...
    does NOT need to implement those endpoints.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/events/{id}/kolaborator:
    get:
      operationId: listEventKolaborator
      tags: [events-kolaborator]
      summary: List kolaborators assigned to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: EventKolaborator list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: assignKolaboratorToEvent
      tags: [events-kolaborator]
      summary: Admin-assign kolaborator to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kolaborator_id, peran]
              properties:
                kolaborator_id:
                  type: string
                peran:
                  type: string
                  enum: [performer, panitia, peserta]
                status_kehadiran:
                  type: string
                  enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
                  default: terdaftar
            example:
              kolaborator_id: kol-0001
              peran: performer
              status_kehadiran: terdaftar
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/events/{id}/kolaborator/{kid}:
    patch:
      operationId: updateEventKolaborator
      tags: [events-kolaborator]
      summary: Update kolaborator assignment in event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: kid
          in: path
          required: true
          schema:
            type: string
          description: EventKolaborator junction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                peran:
                  type: string
                status_kehadiran:
                  type: string
                  enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
            example:
              status_kehadiran: hadir
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: removeEventKolaborator
      tags: [events-kolaborator]
      summary: Remove kolaborator from event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: kid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Kolaborator dihapus dari event
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/kolaborator:
    get:
      operationId: listKolaborator
      tags: [kolaborator-management]
      summary: List all kolaborators (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [aktif, pending, suspended, rejected]
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Kolaborator list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Kolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/kolaborator/{id}:
    get:
      operationId: getKolaborator
      tags: [kolaborator-management]
      summary: Get kolaborator detail (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Kolaborator record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Kolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchKolaborator
      tags: [kolaborator-management]
      summary: Admin update kolaborator record (admin only)
      description: |
        Admin can edit: nama, email, kota, bio, foto_url, cover_url, subsektor, no_hp, internal_notes.
        NOT status (use /status endpoint) — status is exclusively managed via /api/kolaborator/{id}/status.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama:
                  type: string
                email:
                  type: string
                  format: email
                  description: Updates kolaborators.email mirror; backend MUST also update auth.users.
                kota:
                  type: string
                bio:
                  type: string
                no_hp:
                  type: string
                  description: Admin-collected; backend stores in kolaborators.no_hp.
                subsektor:
                  type: array
                  items:
                    type: string
                foto_url:
                  type: string
                  nullable: true
                cover_url:
                  type: string
                  nullable: true
                internal_notes:
                  type: string
            example:
              bio: Desainer tekstil terbaik Banyumas
              no_hp: '081111223344'
              internal_notes: Verified via WhatsApp 2026-04-12.
      responses:
        '200':
          description: Updated kolaborator
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Kolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/kolaborator/{id}/status:
    patch:
      operationId: patchKolaboratorStatus
      tags: [kolaborator-management]
      summary: Update kolaborator status (admin only)
      description: |
        **Dual update required:** Both `app_metadata.status` in Supabase Auth (via Admin SDK)
        AND `status` field in `kolaborators` table must be updated atomically.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [aktif, suspended, rejected]
            example:
              status: aktif
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          status:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/kolaborator/{id}/events:
    get:
      operationId: listKolaboratorEvents
      tags: [kolaborator-management]
      summary: List events a kolaborator has joined (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Events list (event rows the kolaborator has joined)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/KolaboratorEventEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/kolaborator/{id}/portofolio:
    get:
      operationId: listKolaboratorPortofolioAdmin
      tags: [kolaborator-management]
      summary: List kolaborator portfolio items (admin view, admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Karya list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Karya'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/kolaborator/{id}/portofolio/{pid}:
    patch:
      operationId: featureKolaboratorPorto
      tags: [kolaborator-management]
      summary: Toggle featured flag on portfolio item (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pid
          in: path
          required: true
          schema:
            type: string
          description: Portfolio/Karya ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [featured]
              properties:
                featured:
                  type: boolean
            example:
              featured: true
      responses:
        '200':
          description: Updated karya
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Karya'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteKolaboratorPorto
      tags: [kolaborator-management]
      summary: Delete portfolio item (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Item portofolio dihapus
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/kolaborator/{id}/stories:
    get:
      operationId: listKolaboratorStoriesAdmin
      tags: [kolaborator-management]
      summary: List kolaborator stories (admin view, admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Story list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Story'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Petugas Management ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT (RS256). Issued by `supabase.auth.signInWithPassword()` on the frontend.
        Backend verifies using `SUPABASE_JWT_SECRET`.
        Role must be 'admin' OR 'petugas' in `app_metadata.role`.
        Endpoints marked "admin only" additionally reject role 'petugas' with 403.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID or prefixed short ID
      example: evt-0001
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Supabase JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sesi Anda telah berakhir
            data: null
    Forbidden:
      description: Authenticated but role insufficient (e.g. petugas accessing admin-only endpoint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Akses ditolak
            data: null
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sumber daya tidak ditemukan
            data: null
    Conflict:
      description: Conflict (e.g. stand already occupied)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Stand sudah ditempati artisan lain
            data: null
...
  schemas:
    Envelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
          nullable: true
        data:
          description: Actual payload — shape varies per endpoint
    ErrorEnvelope:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        data:
          type: object
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    EventKolaborator:
      type: object
      description: |
        Event-side view: one kolaborator participating in a given event
        (GET /api/events/{id}/kolaborator). nama + subsektor are JOINed from the
        kolaborators table (NOT stored in the event_kolaborators junction).
      required: [id, event_id, kolaborator_id, nama, subsektor, peran, status_kehadiran, assigned_by]
      properties:
        id:
          type: string
          example: evkol-0001
        event_id:
          type: string
          example: evt-0001
        kolaborator_id:
          type: string
          example: kol-0001
        nama:
          type: string
          description: Kolaborator name (JOINed from kolaborators)
          example: Sari Wulandari
        subsektor:
          type: array
          items:
            type: string
          description: BEKRAF subsektor
          example: [Kriya, Fashion]
        peran:
          type: string
          enum: [performer, panitia, peserta]
          example: performer
        status_kehadiran:
          type: string
          enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
          example: terdaftar
        assigned_by:
          type: string
          enum: [admin, self]
          example: admin
    KolaboratorEventEntry:
      type: object
      description: |
        Kolaborator-side view: one event a kolaborator has joined
        (GET /api/kolaborator/{id}/events). nama + date/time/location/status are
        JOINed from the events table; peran/status_kehadiran/assigned_by come
        from the event_kolaborators junction. id is the junction-row id.
      required: [id, event_id, kolaborator_id, nama, tanggal, tanggal_selesai, jam_mulai, jam_selesai, lokasi, status, peran, status_kehadiran, assigned_by, created_at]
      properties:
        id:
          type: string
          description: Junction-row id (event_kolaborators.id)
          example: evkol-0001
        event_id:
          type: string
          example: evt-0001
        kolaborator_id:
          type: string
          example: kol-0001
        nama:
          type: string
          description: Event name (JOINed from events)
          example: Peken Banyumas #7
        tanggal:
          type: string
          format: date
          description: Event start date (WIB local YYYY-MM-DD)
          example: '2026-05-10'
        tanggal_selesai:
          type: string
          format: date
          description: Event end date (WIB local YYYY-MM-DD)
          example: '2026-05-12'
        jam_mulai:
          type: string
          description: WIB 24-hour HH:MM:SS (time serialized with seconds)
          example: '09:00:00'
        jam_selesai:
          type: string
          description: WIB 24-hour HH:MM:SS (time serialized with seconds)
          example: '21:00:00'
        lokasi:
          type: string
          description: Event venue/location name
          example: Alun-alun Purwokerto
        status:
          type: string
          enum: [draft, published, berlangsung, selesai]
          example: published
        peran:
          type: string
          enum: [performer, panitia, peserta]
          example: performer
        status_kehadiran:
          type: string
          enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
          example: terdaftar
        assigned_by:
          type: string
          enum: [admin, self]
          example: admin
        created_at:
          type: string
          format: date-time
          example: '2026-05-01T08:00:00Z'
...
    Kolaborator:
      type: object
      required: [id, slug, nama, email, kota, bio, subsektor, status, tanggal_daftar]
      properties:
        id:
          type: string
          example: kol-0001
        slug:
          type: string
          description: URL-safe slug from nama
          example: sari-wulandari
        nama:
          type: string
          example: Sari Wulandari
        email:
          type: string
          format: email
          example: sari@example.com
        kota:
          type: string
          example: Purwokerto
        bio:
          type: string
          example: Desainer tekstil dengan 7 tahun pengalaman
        foto_url:
          type: string
          nullable: true
          example: https://storage.example.com/foto/kol-0001.jpg
        cover_url:
          type: string
          nullable: true
          example: null
        subsektor:
          type: array
          items:
            type: string
          description: BEKRAF 17 subsectors (NOT kategori_usaha)
          example: [Kriya, Fashion]
        status:
          type: string
          enum: [aktif, pending, suspended, rejected]
          example: aktif
        tanggal_daftar:
          type: string
          format: date
          description: WIB local date YYYY-MM-DD
          example: '2026-01-15'
        total_karya:
          type: integer
          description: Computed by DB trigger
          example: 5
        total_story:
          type: integer
          description: Computed by DB trigger
          example: 3
        total_event:
          type: integer
          description: Computed by DB trigger
          example: 2
        no_hp:
          type: string
          description: |
            **Admin-only** contact phone — admin-collected via Gate Kolaborator edit drawer.
            NEVER returned by /api/public/profiles/{slug} or /api/kolaborator/me.
            Asymmetric by design: only present in gate yaml, not colab yaml.
          example: '08111122331'
        internal_notes:
          type: string
          description: |
            **Admin-only** free-text metadata — admin-collected via Gate Kolaborator edit drawer.
            NEVER returned by /api/public/profiles/{slug} or /api/kolaborator/me.
          example: Verified via WhatsApp 2026-04-12.
...
    Karya:
      type: object
      required: [id, judul, subsektor, deskripsi, tahun, featured, owner_type, owner_id]
      properties:
        id:
          type: string
          example: kar-0001
        judul:
          type: string
          description: Portfolio item title
          example: Batik Geometrik Banyumas
        subsektor:
          type: string
          description: SINGULAR string (NOT array). BEKRAF subsektor.
          example: Kriya
        deskripsi:
          type: string
          example: Motif batik kontemporer terinspirasi arsitektur lokal
        tahun:
          type: integer
          description: Year of creation
          example: 2024
        gambar_url:
          type: string
          nullable: true
          description: Image URL (always _url suffix)
          example: https://storage.example.com/karya/kar-0001.jpg
        featured:
          type: boolean
          description: Admin-toggled featured flag. NOT 'is_featured'.
          example: true
        owner_type:
          type: string
          enum: [artisan, kolaborator]
          description: Polymorphic owner type — points to artisans.id OR kolaborators.id.
          example: kolaborator
        owner_id:
          type: string
          description: UUID of owning artisan or kolaborator (resolves via owner_type).
          example: kol-0001
        owner:
          type: string
          description: Display name of owner (joined from artisans.nama_usaha or kolaborators.nama). Backend computes on read.
          example: Sari Wulandari
        owner_slug:
          type: string
          description: Slug of owner for navigation to /profile/{slug} (joined from artisans.slug or kolaborators.slug).
          example: sari-wulandari
        created_at:
          type: string
          format: date-time
          description: UTC ISO 8601, set on INSERT
          example: '2026-04-01T08:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: UTC ISO 8601, auto-set by set_updated_at trigger on UPDATE
          example: '2026-04-15T10:30:00.000Z'
    Story:
      type: object
      required: [id, konten, tags, like_count, status, created_at]
      properties:
        id:
          type: string
          example: sto-0001
        konten:
          type: string
          description: Story text content
          example: Mengikuti pameran di Jakarta minggu lalu
        media_url:
          type: string
          nullable: true
          description: Optional attached media URL
          example: null
        tags:
          type: array
          items:
            type: string
          description: Free-form tags
          example: [pameran, kriya]
        like_count:
          type: integer
          description: Number of likes (computed)
          example: 12
        status:
          type: string
          enum: [aktif, dihapus, disembunyikan]
          example: aktif
        created_at:
          type: string
          format: date-time
          description: UTC ISO 8601
          example: '2026-04-01T08:00:00.000Z'
...